fix: new header layout, moved export to details page
