+++ src/app/screens/Accounts/Detail/index.tsx
@@ -1,3 +1,4 @@
+import { CrossIcon } from "@bitcoin-design/bitcoin-icons-react/filled";
 import Button from "@components/Button";
 import Container from "@components/Container";
 import Loading from "@components/Loading";
@@ -7,11 +8,14 @@ import dayjs from "dayjs";
 import relativeTime from "dayjs/plugin/relativeTime";
 import type { FormEvent } from "react";
 import { useCallback, useEffect, useRef, useState } from "react";
-import { useTranslation } from "react-i18next";
+import { Trans, useTranslation } from "react-i18next";
+import Modal from "react-modal";
+import QRCode from "react-qr-code";
 import { Link, useNavigate, useParams } from "react-router-dom";
 import { toast } from "react-toastify";
 import Alert from "~/app/components/Alert";
 import Badge from "~/app/components/Badge";
+import Hyperlink from "~/app/components/Hyperlink";
 import InputCopyButton from "~/app/components/InputCopyButton";
 import MenuDivider from "~/app/components/Menu/MenuDivider";
 import Select from "~/app/components/form/Select";
@@ -20,6 +24,7 @@ import { useAccount } from "~/app/context/AccountContext";
 import { useAccounts } from "~/app/context/AccountsContext";
 import { useSettings } from "~/app/context/SettingsContext";
 import api, { GetAccountRes } from "~/common/lib/api";
+import msg from "~/common/lib/msg";
 import nostr from "~/common/lib/nostr";
 import type { Account, BitcoinNetworkType } from "~/types";
 
@@ -36,15 +41,37 @@ function AccountDetail() {
   const { isLoading: isLoadingSettings } = useSettings();
 
   const hasFetchedData = useRef(false);
-  const [account, setAccount] = useState<GetAccountRes | null>(null);
   const { id } = useParams() as { id: string };
   const navigate = useNavigate();
 
+  const [account, setAccount] = useState<GetAccountRes | null>(null);
   const [accountName, setAccountName] = useState("");
-
   const [hasMnemonic, setHasMnemonic] = useState(false);
   const [nostrPublicKey, setNostrPublicKey] = useState("");
   const [hasImportedNostrKey, setHasImportedNostrKey] = useState(false);
+
+  const [exportLoading, setExportLoading] = useState(false);
+  const [exportModalIsOpen, setExportModalIsOpen] = useState(false);
+  const [lndHubData, setLndHubData] = useState({
+    login: "",
+    password: "",
+    url: "",
+    lnAddress: "",
+  });
+  async function exportAccount({ id, name }: AccountAction) {
+    setExportLoading(true);
+    setExportModalIsOpen(true);
+    setLndHubData(
+      await msg.request("accountDecryptedDetails", {
+        name,
+        id,
+      })
+    );
+    setExportLoading(false);
+  }
+  function closeExportModal() {
+    setExportModalIsOpen(false);
+  }
 
   const fetchData = useCallback(async () => {
     try {
@@ -147,7 +174,7 @@ function AccountDetail() {
         </div>
 
         <div>
-          <div className="shadow bg-white sm:rounded-md sm:overflow-hidden px-6 py-2 divide-y divide-black/10 dark:divide-white/10 dark:bg-surface-02dp">
+          <div className="shadow bg-white sm:rounded-md sm:overflow-hidden p-6 dark:bg-surface-02dp flex flex-col gap-4">
             <form
               onSubmit={(e: FormEvent) => {
                 e.preventDefault();
@@ -159,7 +186,7 @@ function AccountDetail() {
                 updatedAccount.name = accountName;
                 setAccount(updatedAccount);
               }}
-              className="my-4 flex justify-between items-end"
+              className="flex justify-between items-end"
             >
               <div className="w-7/12">
                 <TextField
@@ -184,6 +211,111 @@ function AccountDetail() {
                 />
               </div>
             </form>
+            {account.connectorType == "lndhub" && (
+              <>
+                <MenuDivider />
+                <div className="flex justify-between items-end">
+                  <div className="w-9/12">
+                    <p className="text-gray-900 dark:text-white font-medium">
+                      {t("export.title")}
+                    </p>
+                    <p className="text-gray-500 text-sm dark:text-neutral-500">
+                      <Trans
+                        i18nKey={"export.description"}
+                        t={t}
+                        components={[
+                          // eslint-disable-next-line react/jsx-key
+                          <Hyperlink
+                            href="https://bluewallet.io"
+                            target="_blank"
+                            rel="noopener nofollow"
+                          />,
+                          // eslint-disable-next-line react/jsx-key
+                          <Hyperlink
+                            href="https://zeusln.app"
+                            target="_blank"
+                            rel="noopener nofollow"
+                          />,
+                        ]}
+                      />
+                    </p>
+                  </div>
+
+                  <div className="w-1/5 flex-none">
+                    {exportAccount && account.connectorType === "lndhub" && (
+                      <>
+                        <Button
+                          label={tCommon("actions.connect")}
+                          fullWidth
+                          onClick={() =>
+                            exportAccount({
+                              id: account.id,
+                              name: account.name,
+                            })
+                          }
+                        />
+                      </>
+                    )}
+                  </div>
+                  <Modal
+                    ariaHideApp={false}
+                    closeTimeoutMS={200}
+                    isOpen={exportModalIsOpen}
+                    onRequestClose={closeExportModal}
+                    contentLabel={t("export.screen_reader")}
+                    overlayClassName="bg-black bg-opacity-50 fixed inset-0 flex justify-center items-center p-5"
+                    className="rounded-lg bg-white dark:bg-surface-02dp w-full max-w-md overflow-hidden"
+                  >
+                    <div className="p-5 flex justify-between">
+                      <h2 className="text-2xl font-bold dark:text-white">
+                        {t("export.title")}
+                      </h2>
+                      <button onClick={closeExportModal}>
+                        <CrossIcon className="w-6 h-6 dark:text-white" />
+                      </button>
+                    </div>
+
+                    {exportLoading && (
+                      <div className="p-5 flex justify-center items-center space-x-2 dark:text-white">
+                        <Loading />
+                        <span>{t("export.waiting")}</span>
+                      </div>
+                    )}
+                    {!exportLoading && (
+                      <div className="p-5">
+                        <div className="flex flex-col gap-4 justify-center items-center dark:text-white">
+                          <p>{t("export.scan_qr")}</p>
+                          <QRCode
+                            value={`lndhub://${lndHubData.login}:${lndHubData.password}@${lndHubData.url}/`}
+                            level="M"
+                            size={256}
+                          />
+                          <div className="w-full">
+                            <TextField
+                              id="uri"
+                              label={t("export.export_uri")}
+                              type="text"
+                              readOnly
+                              value={`lndhub://${lndHubData.login}:${lndHubData.password}@${lndHubData.url}/`}
+                            />
+                          </div>
+                          {lndHubData.lnAddress && (
+                            <div className="w-full dark:text-white">
+                              <p className="font-medium">
+                                {t("export.your_ln_address")}
+                              </p>
+                              {lndHubData.lnAddress && (
+                                <p>{lndHubData.lnAddress}</p>
+                              )}
+                            </div>
+                          )}
+                        </div>
+                      </div>
+                    )}
+                  </Modal>
+                </div>
+              </>
+            )}
           </div>
 
           <h2 className="text-2xl mt-12 font-bold dark:text-white">
